LayoutTests/js/regress/JSONP-negative-0-expected.txt

JSRegress/JSONP-negative-0

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.is(-0,x) is true
PASS no exception thrown
PASS successfullyParsed is true

TEST COMPLETE
